Update GitHub action to support cache and build (#151)
The Zotero bibliography.json file is cached.  On a push to the main
directory, the cache is checked to see if it is out dated, if so
it is updated prior to deploying the updated web site.  In addition,
a schedule job runs periodically, checking to see if the bibliography
cache is out of date.  If so, it is refreshed and a new version is
pulled down and the web page is redeployed with the updated
bibliography.