Add Title to hugo.yaml and modify navbar to use it
navbar.html by defalut puts the title on every page.  This update
modifies it to not put the title on the home page.  The title
already is a central part of the text within the page and having
it on the home page is redundant.

Defining the title in the config comes in useful for other Hugo
templates that want to use the title of the website.