Add support for  RSS Feed (#258)
* Add Title to hugo.yaml and modify navbar to use it

navbar.html by defalut puts the title on every page.  This update
modifies it to not put the title on the home page.  The title
already is a central part of the text within the page and having
it on the home page is redundant.

Defining the title in the config comes in useful for other Hugo
templates that want to use the title of the website.

* Add RSS Feed to History, Software and Project pages.

RSS Feed generates feed containing all subpages from these
sections.

* Add rss Feed to Footer

Update rssFeed url value to project/status page

Add frontmatter and enable git info

These changes allow us to get the last modified date for pages.

Update page-meta-lastmod file to not show git commit info

* Update config files to rename RSS Feed to rss.xml

* Update title and description lines

Title and Description are set to Updates from the Medley Interlisp
Project.