Update css to make navbar more responsive
Updates to make the navigation bar more responsive
to the browser window size.

Change 1: Drop the title, The Medley Interlisp Project, when
the size of the screen shrinks to a point where there is no space
between the icons and the title.

Change 2: On the home page, when the icons and search bar have
taken up all the horizontal space, drop the search bar to a
second line.

Change 3: On all other pages, drop the search bar, we have the
left search bar available.