Adding a CCECE 2025 paper section to the News and Status Reports page (#283)
* Initial commit

Upload IEEE CCEECE 2025 paper preprint to link from the News and Status Reports page.

Signed-off-by: Paolo Amoroso <paolo.amoroso@gmail.com>

* Add CCECE 2025 section

Add a section on the CCEECE 2025 talk by Eleanor Young with a link to the preprint. This addresses Medley issue #2175.

Signed-off-by: Paolo Amoroso <info@paoloamoroso.com>

* Rephrase talk description/preprint, drop CCECE link

Rephrase the reference to the CCECE talk. Remove the link to the CCECE technical program. Rephrase the reference to the downloadable preprint for smoother flow.

These changes address feedback by Bill Stumbo.

Signed-off-by: Paolo Amoroso <info@paoloamoroso.com>

---------

Signed-off-by: Paolo Amoroso <paolo.amoroso@gmail.com>
Signed-off-by: Paolo Amoroso <info@paoloamoroso.com>