250715 link checker updates (#288)
* Add Release Notes

Add section Release Notes moved here from the News and Status Reports page.

Signed-off-by: Paolo Amoroso <info@paoloamoroso.com>

* Rearrange and delete irrelevant sections

Rearrange the entries in reverse chronological order. Delete sections Build Notes (moved to the Install and Run page) and Closed Issues which are not relevant here.

Signed-off-by: Paolo Amoroso <info@paoloamoroso.com>

* Add section Books and Publications

Add to the page What people are saying the new section Books and Publications with an entry for the book Hypertext.

Signed-off-by: Paolo Amoroso <info@paoloamoroso.com>

* Fix 404 errors and redirections found by W3C link checker.

---------

Signed-off-by: Paolo Amoroso <info@paoloamoroso.com>
Co-authored-by: Paolo Amoroso <info@paoloamoroso.com>