Clean up page Get Involved
Revise the Get Involved page as per issue #2217 to clarify some link anchors; identify the GitHub repo of Interlisp.github.io; remove the Mastodon and Bluesky links from section Chat and move the project's email details to section EMail; rename section Chat to Online communities and add a call for volunteers to monitor those platforms; and shorten the Weekly meetings heading by moving part of it to the section text. Point out that we have official accounts on Mastodon and Bluseky and no presence on Facebook.

Signed-off-by: Paolo Amoroso <info@paoloamoroso.com>