Fixes rss feed.  Changes to _index.md are now included.
The code was including all the pages within a section but not the
section index, _index.md.  This was fixed by specifically adding
the section index page:

```
  {{- $pages = $pages | union (slice .) }}
```

and then adding the pages within the section:

```
{{- $pages = $pages | union (slice .) }}
```

In addition, the pages are ordered via Publish Date before the
resulting feed is build.