Initial implementation of list of all bibliography entries
Primitive first pass.  Provides a chronological list of all
bibliography entires providing titles, authors, date and an
abbreviated start of the abstract (limted to 24 words). In
addition, a link is enclosed that goes to the markdown file
that represents the bibliography entry.

Currently, there is nothing implemented to show the contents of
the md file.

There is no bibliography css implemented to format the list items.

This provides an initial proof of concept.  A simple framework that
can be extended and styled to get us where we want to go.