Add lastmod, last modified date of the bibliographic entry (#302)
lastmod is used in the sitemap to help identify files that have changed
and should be recrawled.  Writing lastmod into the markdown file's
YAML will ensure it takes presidence over the current date field,
which represents the publication date of the item, not the last
time the bibliographic entry was modified.