Bs7 citations (#305)
* Initial attempt to build citation modal dialog box.

Citation link, generates modal dialog box whose content is the
results of a call to Zotero to get the citation.  Currently
using apa citation, there are many other options...

* Update Citation modal to include both rendered text and html

Allow the user to copy either the formatted 'pretty' text or
the raw html that creates the entry.

* Fix reference to prettier js files in script.

* Convert 'View on Zotero' and 'Citation' into buttons.

Move citations button to bottom of page with Zotero button.

* Move Zotero group id from hardcoded to a site parameter

* Fix button styles and clean up Hugo operation syntax.

In the process I discovered a couple minor errors in the
layout of the bibliography and fixed them.