Bs10 filter by year (#312)
* Initial implementation of adding filter by year

* Update dates to simple expressions

* Add new version of date selection

This version replaces using tags with a text window that allows
entering single dates, a relationship (<,>, etc.) and a date or
the phrase `date between date` allowing a selection of dates.

* Change 'between' to 'to'.  Fix comparison to use start of year.

* Updates to fix comparison operators

* Remove unneeded JSON creation.  More comparison operator cleanup.

* Add tooltip, reformatting leverage bootstrap, less custom css

* Stop date field for erasing on CR

* Overhaul to focus on just years.

Input is limited to just years and comparison operators.
Changed so processing of input occurs on CR.  Include
validation of input.  Revised Tooltip to hide once user
starts entering input.